+++ resources/js/humhub.wiki.Form.js
@@ -1,0 +1,34 @@
+humhub.module('wiki.Form', function(module, require, $) {
+    var Widget = require('ui.widget').Widget;
+    var wikiView = require('wiki');
+
+    /**
+     * This widget represents the wiki form
+     */
+    var Form = Widget.extend();
+
+    Form.prototype.init = function() {
+        var that = this;
+        this.getRichtext().on('init', function() {
+            wikiView.registerStickyElement(that.getRichtextMenu(), that.getRichtext());
+        });
+    };
+
+    Form.prototype.getRichtextMenu = function() {
+        if(!this.$menu) {
+            this.$menu = this.$.find('.ProseMirror-menubar');
+        }
+
+        return this.$menu;
+    };
+
+    Form.prototype.getRichtext = function() {
+        if(!this.$richtext) {
+            this.$richtext = this.$.find('.ProsemirrorEditor');
+        }
+
+        return this.$richtext;
+    };
+
+    module.export = Form;
+});